backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
prometheus-client
httpx
alembic
email-validator
python-dotenv
slowapi
structlog
tenacity
